+++ utils/go.mod
@@ -2,4 +2,9 @@ module utils
 
 go 1.16
 
-require github.com/spf13/viper v1.7.1
+require (
+	github.com/spf13/viper v1.7.1
+	models v0.0.0-00010101000000-000000000000
+)
+
+replace models => ../models
